Move component styles to generic file naming
