+++ packages/cfpb-design-system/src/elements/cfpb-button/cfpb-button.scss
@@ -1,7 +1,7 @@
 @use 'sass:math';
 @use './vars' as *;
 @use '@cfpb/cfpb-design-system/src/elements/abstracts' as *;
-@use '@cfpb/cfpb-design-system/src/elements/cfpb-icon-text/cfpb-icon-text.component'
+@use '@cfpb/cfpb-design-system/src/elements/cfpb-icon-text/styles.component'
   as *;
 
 :host {
