+++ packages/cfpb-design-system/src/elements/cfpb-checkbox-icon/styles.component.scss
