+++ packages/cfpb-design-system/src/elements/cfpb-expandable/styles.component.scss
@@ -2,7 +2,7 @@
 @use '@cfpb/cfpb-design-system/src/elements/abstracts' as *;
 @use '@cfpb/cfpb-design-system/src/elements/base' as *;
 @use '@cfpb/cfpb-design-system/src/utilities' as *;
-@use '../cfpb-utilities/transition/transition.scss' as *;
+@use '../cfpb-utilities/transition/transition.css' as *;
 
 :host {
   // Theme
