+++ packages/cfpb-design-system/src/elements/cfpb-file-upload/index.js
@@ -1,6 +1,6 @@
 import { LitElement, html, css, unsafeCSS } from 'lit';
 import { ref, createRef } from 'lit/directives/ref.js';
-import styles from './cfpb-file-upload.component.scss?inline';
+import styles from './styles.component.scss?inline';
 import { CfpbButton } from '../cfpb-button';
 
 /**
