+++ packages/cfpb-design-system/src/elements/cfpb-file-upload/styles.component.scss
