+++ packages/cfpb-design-system/src/elements/cfpb-form-choice/index.js
@@ -1,7 +1,7 @@
 import { LitElement, html, css, unsafeCSS } from 'lit';
 import { classMap } from 'lit/directives/class-map.js';
 import { ref, createRef } from 'lit/directives/ref.js';
-import styles from './cfpb-form-choice.component.scss?inline';
+import styles from './styles.component.scss?inline';
 import { CfpbCheckboxIcon } from '../cfpb-checkbox-icon';
 
 // The validation states are error, warning, or success.
