+++ packages/cfpb-design-system/src/elements/cfpb-form-search-input/styles.component.scss
