+++ packages/cfpb-design-system/src/elements/cfpb-form-search/styles.component.scss
