+++ packages/cfpb-design-system/src/elements/cfpb-icon-text/styles.component.scss
