+++ packages/cfpb-design-system/src/elements/cfpb-label/styles.component.scss
