+++ packages/cfpb-design-system/src/elements/cfpb-listbox-item/styles.component.scss
