+++ packages/cfpb-design-system/src/elements/cfpb-listbox/styles.component.scss
