+++ packages/cfpb-design-system/src/elements/cfpb-pagination/styles.component.scss
