+++ packages/cfpb-design-system/src/elements/cfpb-select/styles.component.scss
@@ -1,7 +1,7 @@
 @use '@cfpb/cfpb-design-system/src/elements/base' as *;
 @use '@cfpb/cfpb-design-system/src/utilities' as *;
 @use '@cfpb/cfpb-design-system/src/components/cfpb-icons/icon';
-@use '../cfpb-utilities/transition/transition.scss' as *;
+@use '../cfpb-utilities/transition/transition.css' as *;
 @use '../../utilities/functions' as *;
 
 :host {
