+++ packages/cfpb-design-system/src/elements/cfpb-tagline/styles.component.scss
