+++ packages/cfpb-design-system/src/elements/cfpb-utilities/transition/transition.css
@@ -9,9 +9,9 @@
   transition-duration: 0s !important;
 }
 
-//
-// Utility classes for moving an element using transform translate values.
-//
+/**
+ * Utility classes for moving an element using transform translate values.
+ */
 
 .u-move-transition {
   transition: transform 0.25s ease-out;
@@ -25,7 +25,7 @@
   transform: translate3d(-100%, 0, 0);
 }
 
-// TODO: Look into adding a mixin for movement multiples.
+/* TODO: Look into adding a mixin for movement multiples. */
 .u-move-left-2x {
   transform: translate3d(-200%, 0, 0);
 }
@@ -42,9 +42,9 @@
   transform: translate3d(0, -100%, 0);
 }
 
-//
-// Utility classes for setting an element's opacity.
-//
+/**
+ * Utility classes for setting an element's opacity.
+ */
 
 .u-alpha-transition {
   transition: opacity 0.25s linear;
@@ -58,15 +58,15 @@
   opacity: 0;
 }
 
-//
-// Utility classes for setting an element's height.
-//
+/**
+ * Utility classes for setting an element's height.
+ */
 
 .u-max-height-transition {
   overflow: hidden;
   contain: paint;
 
-  // Duration is set here, but it is actually overridden in the JavaScript.
+  /* Duration is set here, but it is actually overridden in the JavaScript. */
   transition: max-height 0.2s ease-out;
 }
