+++ packages/cfpb-design-system/src/index.scss
@@ -74,4 +74,4 @@
 @forward 'elements/cfpb-button/vars';
 
 // cfpb-icon-text
-@forward 'elements/cfpb-icon-text/cfpb-icon-text.component';
+@forward 'elements/cfpb-icon-text/styles.component';
